+++ common/src/main/scala/com/stratio/crossdata/common/serializers/SQLResultSerializer.scala
@@ -13,19 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.crossdata.serializers
+package com.stratio.crossdata.common.serializers
 
 import com.stratio.crossdata.common.result.{ErrorSQLResult, SQLResult, SuccessfulSQLResult}
 import org.apache.spark.sql.Row
+import org.apache.spark.sql.crossdata.serializers.StructTypeSerializer
 import org.apache.spark.sql.types.StructType
 import org.json4s.JsonDSL._
 import org.json4s.{CustomSerializer, Extraction, _}
 
 object SQLResultSerializer extends CustomSerializer[SQLResult]( format => (
   {
-    case JObject(JField("successfulResult", JBool(false))::JField("error", JString(message))::Nil) =>
+    case JObject(JField("successfulResult", JBool(false))::JField("error", JString(message))::_) =>
       ErrorSQLResult(message)
-    case JObject(JField("successfulResult", JBool(true))::JField("schema", json_schema)::JField("rows", JArray(json_rows)) ::Nil) =>
+    case JObject(JField("successfulResult", JBool(true))::JField("schema", json_schema)::JField("rows", JArray(json_rows)) ::_) =>
 
       val schema = {
         implicit val formats = DefaultFormats + StructTypeSerializer