+++ mongodb/src/main/scala/com/stratio/crossdata/connector/mongodb/MongoQueryProcessor.scala
@@ -27,9 +27,9 @@ import com.stratio.datasource.mongodb.schema.MongodbRowConverter._
 import com.stratio.datasource.util.Config
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.planning.PhysicalOperation
-import org.apache.spark.sql.catalyst.plans.logical.{Limit => LogicalLimit, LogicalPlan}
+import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, Limit => LogicalLimit}
 import org.apache.spark.sql.{Row, sources}
-import org.apache.spark.sql.sources.CatalystToCrossdataAdapter.{BaseLogicalPlan, FilterReport, ProjectReport, SimpleLogicalPlan}
+import org.apache.spark.sql.sources.CatalystToCrossdataAdapter.{BaseLogicalPlan, FilterReport, ProjectReport, SimpleLogicalPlan, CrossdataExecutionPlan}
 import org.apache.spark.sql.sources.{CatalystToCrossdataAdapter, Filter => SourceFilter}
 import org.apache.spark.sql.types.StructType
 
@@ -201,9 +201,9 @@ class MongoQueryProcessor(logicalPlan: LogicalPlan, config: Config, schemaProvid
 
       case PhysicalOperation(projectList, filterList, _) =>
         CatalystToCrossdataAdapter.getConnectorLogicalPlan(logicalPlan, projectList, filterList) match {
-          case (_, ProjectReport(exprIgnored), FilterReport(filtersIgnored, _)) if filtersIgnored.nonEmpty || exprIgnored.nonEmpty =>
+          case CrossdataExecutionPlan(_, ProjectReport(exprIgnored), FilterReport(filtersIgnored, _)) if filtersIgnored.nonEmpty || exprIgnored.nonEmpty =>
             None
-          case (basePlan: SimpleLogicalPlan, _, _) =>
+          case CrossdataExecutionPlan(basePlan: SimpleLogicalPlan, _, _) =>
             Some(basePlan)
           case _ => ??? // TODO
         }