+++ server/src/main/scala/com/stratio/crossdata/server/actors/ServerActor.scala
@@ -117,7 +117,10 @@ class ServerActor(cluster: Cluster, sessionProvider: XDSessionProvider)
 
         case Failure(error) =>
           logger.warn(s"Received message with an unknown sessionId $id", error)
-          sender ! ErrorSQLResult(s"Unable to recover the session ${session.id}. Cause: ${error.getMessage}")
+          sender ! SQLReply(
+            sqlCommand.requestId,
+            ErrorSQLResult(s"Unable to recover the session ${session.id}. Cause: ${error.getMessage}")
+          )
       }
 
 
@@ -177,7 +180,7 @@ class ServerActor(cluster: Cluster, sessionProvider: XDSessionProvider)
     case sc@CommandEnvelope(_: OpenSessionCommand, session, _) =>
       val open = sessionProvider.newSession(session.id) match {
         case Success(_) =>
-          logger.debug(s"new session with sessionID=${session.id} has been created")
+          logger.info(s"new session with sessionID=${session.id} has been created")
           true
         case Failure(error) =>
           logger.error(s"failure while creating the session with sessionID=${session.id}")